Refactor: Modularize keyboard layout mappings
This commit refactors the keyboard layout mapping to be more modular
and extensible. The hardcoded hebrew-to-qwerty mapping has been
replaced with a generic Keymap structure that can support multiple
languages. A new file, src/keymap.go, has been introduced to store
the mappings, and the pattern matching logic has been updated to use
the new structure.

This change makes it easier to add support for new keyboard layouts
in the future.